add tests for keydown on Time picker
