+++ packages/components/src/DateTimePickers/InputDateTimePicker/InputDateTimePicker.component.test.js
@@ -1,29 +1,10 @@
 import React from 'react';
 import { mount } from 'enzyme';
-import keycode from 'keycode';
 
 import InputDateTimePicker from './InputDateTimePicker.component';
 import Manager from '../DateTime/Manager';
 
 describe('InputDateTimePicker', () => {
-	describe('keydown', () => {
-		it('should focus on calendar day if it is open with input DOWN', () => {
-			// given
-			const wrapper = mount(<InputDateTimePicker id="my-picker" />);
-			wrapper.simulate('focus');
-			const event = { keyCode: keycode.codes.down };
-
-			// when
-			wrapper
-				.find('input')
-				.first()
-				.simulate('keydown', event);
-
-			// then
-			expect(document.activeElement.classList.contains('tc-date-picker-day')).toBe(true);
-		});
-	});
-
 	describe('onChange', () => {
 		it('should trigger props.onChange', () => {
 			// given
